+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.610] - 2022-05-12
+## Changed
+- BlazorDeviceControl
+- DataCore
+- ScalesUI
+
+## [0.6.605] - 2022-05-11
+## Changed
+- BlazorDeviceControl
+- DataCore
+- ScalesUI
+
 ## [0.6.580] - 2022-05-05
 ## Changed
 - BlazorDeviceControl